backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.7
pillow==10.3.0
numpy==1.24.4
scikit-learn==1.3.2
tensorflow==2.15.0
mediapipe
uvicorn
